LayoutTests/platform/blackberry/media/short-media-repeats-correctly-expected.txt

Test that a very short audio file does not repeat unintentionally.

EVENT(canplaythrough)
EVENT(playing)
EVENT(ended)
EVENT(playing)
EVENT(ended)
EVENT(progress)
EVENT(playing)
EVENT(ended)
EVENT(progress)
EVENT(playing)
EVENT(ended)
EVENT(progress)
EVENT(playing)
EVENT(ended)
EVENT(progress)
EVENT(playing)
EVENT(ended)
EVENT(progress)
EXPECTED(progressCount == '5') OK
END OF TEST
